warning shows only once by repeated importation
